feat: DNG support in open modal, drag-and-drop, and File → Open
Extract importDngFile to src/io/dng.ts (shared module). Wire up DNG
routing in all file open paths:
- NewDocumentModal "Open File" button accepts .dng
- File → Open accepts .dng
- Drag-and-drop accepts .dng files
- Pre-document open handler routes .dng to the WASM decoder
