+++ src/app/MenuBar/menus/file-menu.ts
@@ -10,7 +10,6 @@ import {
   compositeForExport,
   exportPng16,
   getCompositeSize,
-  decodeAndUploadDng,
 } from '../../../engine-wasm/wasm-bridge';
 import type { MenuDef } from './types';
 import { exportPsdFile, importPsdFile } from '../../../io/psd';
@@ -19,56 +18,9 @@ import { describeError, notifyError } from '../../notifications-store';
 // Re-export so existing callers (App.tsx, e2e tests) keep working.
 export { importPsdFile, exportPsdFile };
 
-import { getEngine as getEngineRef } from '../../../engine-wasm/engine-state';
-import { resetTrackedState } from '../../../engine-wasm/engine-sync';
+import { importDngFile } from '../../../io/dng';
 
 const METADATA_NOTE = 'Made with Lopsy — http://lopsy.art';
-
-function importDngFile(data: Uint8Array, name: string): void {
-  // DNG decode happens entirely in WASM. We need to know the dimensions first
-  // so we can create a document at the right size.
-  // Strategy: create a temporary 1x1 doc, decode+upload, then patch dimensions.
-
-  const edStore = useEditorStore.getState();
-  edStore.createDocument(1, 1, false);
-
-  const engine = getEngineRef();
-  if (!engine) {
-    notifyError('Engine not ready');
-    return;
-  }
-
-  const activeLayerId = useEditorStore.getState().document.activeLayerId;
-  if (!activeLayerId) {
-    notifyError('No active layer');
-    return;
-  }
-
-  const dims = decodeAndUploadDng(engine, activeLayerId, data);
-  const width = dims[0] ?? 0;
-  const height = dims[1] ?? 0;
-
-  if (width === 0 || height === 0) {
-    notifyError('DNG decode returned empty image');
-    return;
-  }
-
-  // Patch document and layer dimensions to match the decoded DNG
-  useEditorStore.setState((s) => {
-    const layers = s.document.layers.map((l) => {
-      if (l.id === activeLayerId && l.type === 'raster') {
-        return { ...l, width, height, name };
-      }
-      return l;
-    });
-    return {
-      document: { ...s.document, width, height, layers, name },
-    };
-  });
-
-  resetTrackedState(engine);
-  useEditorStore.getState().fitToView();
-}
 
 function confirmIfDirty(): boolean {
   if (!useEditorStore.getState().isDirty) return true;
