+++ src/components/NewDocumentModal/NewDocumentModal.tsx
@@ -126,7 +126,7 @@ export function NewDocumentModal({ onCreateDocument, onOpenFile, onPasteClipboar
   const handleOpenFile = useCallback(() => {
     const input = document.createElement('input');
     input.type = 'file';
-    input.accept = 'image/*,.psd';
+    input.accept = 'image/*,.psd,.dng';
     input.onchange = () => {
       const file = input.files?.[0];
       if (file) onOpenFile(file);
